Module aws_sdk_auditmanager::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AssociateAssessmentReportEvidenceFolder

Fluent builder constructing a request to AssociateAssessmentReportEvidenceFolder.

BatchAssociateAssessmentReportEvidence

Fluent builder constructing a request to BatchAssociateAssessmentReportEvidence.

BatchCreateDelegationByAssessment

Fluent builder constructing a request to BatchCreateDelegationByAssessment.

BatchDeleteDelegationByAssessment

Fluent builder constructing a request to BatchDeleteDelegationByAssessment.

BatchDisassociateAssessmentReportEvidence

Fluent builder constructing a request to BatchDisassociateAssessmentReportEvidence.

BatchImportEvidenceToAssessmentControl

Fluent builder constructing a request to BatchImportEvidenceToAssessmentControl.

CreateAssessment

Fluent builder constructing a request to CreateAssessment.

CreateAssessmentFramework

Fluent builder constructing a request to CreateAssessmentFramework.

CreateAssessmentReport

Fluent builder constructing a request to CreateAssessmentReport.

CreateControl

Fluent builder constructing a request to CreateControl.

DeleteAssessment

Fluent builder constructing a request to DeleteAssessment.

DeleteAssessmentFramework

Fluent builder constructing a request to DeleteAssessmentFramework.

DeleteAssessmentFrameworkShare

Fluent builder constructing a request to DeleteAssessmentFrameworkShare.

DeleteAssessmentReport

Fluent builder constructing a request to DeleteAssessmentReport.

DeleteControl

Fluent builder constructing a request to DeleteControl.

DeregisterAccount

Fluent builder constructing a request to DeregisterAccount.

DeregisterOrganizationAdminAccount

Fluent builder constructing a request to DeregisterOrganizationAdminAccount.

DisassociateAssessmentReportEvidenceFolder

Fluent builder constructing a request to DisassociateAssessmentReportEvidenceFolder.

GetAccountStatus

Fluent builder constructing a request to GetAccountStatus.

GetAssessment

Fluent builder constructing a request to GetAssessment.

GetAssessmentFramework

Fluent builder constructing a request to GetAssessmentFramework.

GetAssessmentReportUrl

Fluent builder constructing a request to GetAssessmentReportUrl.

GetChangeLogs

Fluent builder constructing a request to GetChangeLogs.

GetControl

Fluent builder constructing a request to GetControl.

GetDelegations

Fluent builder constructing a request to GetDelegations.

GetEvidence

Fluent builder constructing a request to GetEvidence.

GetEvidenceByEvidenceFolder

Fluent builder constructing a request to GetEvidenceByEvidenceFolder.

GetEvidenceFolder

Fluent builder constructing a request to GetEvidenceFolder.

GetEvidenceFoldersByAssessment

Fluent builder constructing a request to GetEvidenceFoldersByAssessment.

GetEvidenceFoldersByAssessmentControl

Fluent builder constructing a request to GetEvidenceFoldersByAssessmentControl.

GetOrganizationAdminAccount

Fluent builder constructing a request to GetOrganizationAdminAccount.

GetServicesInScope

Fluent builder constructing a request to GetServicesInScope.

GetSettings

Fluent builder constructing a request to GetSettings.

ListAssessmentFrameworkShareRequests

Fluent builder constructing a request to ListAssessmentFrameworkShareRequests.

ListAssessmentFrameworks

Fluent builder constructing a request to ListAssessmentFrameworks.

ListAssessmentReports

Fluent builder constructing a request to ListAssessmentReports.

ListAssessments

Fluent builder constructing a request to ListAssessments.

ListControls

Fluent builder constructing a request to ListControls.

ListKeywordsForDataSource

Fluent builder constructing a request to ListKeywordsForDataSource.

ListNotifications

Fluent builder constructing a request to ListNotifications.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

RegisterAccount

Fluent builder constructing a request to RegisterAccount.

RegisterOrganizationAdminAccount

Fluent builder constructing a request to RegisterOrganizationAdminAccount.

StartAssessmentFrameworkShare

Fluent builder constructing a request to StartAssessmentFrameworkShare.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateAssessment

Fluent builder constructing a request to UpdateAssessment.

UpdateAssessmentControl

Fluent builder constructing a request to UpdateAssessmentControl.

UpdateAssessmentControlSetStatus

Fluent builder constructing a request to UpdateAssessmentControlSetStatus.

UpdateAssessmentFramework

Fluent builder constructing a request to UpdateAssessmentFramework.

UpdateAssessmentFrameworkShare

Fluent builder constructing a request to UpdateAssessmentFrameworkShare.

UpdateAssessmentStatus

Fluent builder constructing a request to UpdateAssessmentStatus.

UpdateControl

Fluent builder constructing a request to UpdateControl.

UpdateSettings

Fluent builder constructing a request to UpdateSettings.

ValidateAssessmentReportIntegrity

Fluent builder constructing a request to ValidateAssessmentReportIntegrity.