Module aws_sdk_auditmanager::model

Data structures used by operation inputs/outputs.

Modules

assessment

See Assessment

assessment_control

See AssessmentControl

assessment_control_set

See AssessmentControlSet

assessment_evidence_folder

See AssessmentEvidenceFolder

assessment_framework

See AssessmentFramework

assessment_framework_metadata

See AssessmentFrameworkMetadata

assessment_framework_share_request

See AssessmentFrameworkShareRequest

assessment_metadata

See AssessmentMetadata

assessment_metadata_item

See AssessmentMetadataItem

assessment_report

See AssessmentReport

assessment_report_evidence_error

See AssessmentReportEvidenceError

assessment_report_metadata

See AssessmentReportMetadata

assessment_reports_destination

See AssessmentReportsDestination

aws_account

See AwsAccount

aws_service

See AwsService

batch_create_delegation_by_assessment_error

See BatchCreateDelegationByAssessmentError

batch_delete_delegation_by_assessment_error

See BatchDeleteDelegationByAssessmentError

batch_import_evidence_to_assessment_control_error

See BatchImportEvidenceToAssessmentControlError

change_log

See ChangeLog

control

See Control

control_comment

See ControlComment

control_mapping_source

See ControlMappingSource

control_metadata

See ControlMetadata

control_set

See ControlSet

create_assessment_framework_control

See CreateAssessmentFrameworkControl

create_assessment_framework_control_set

See CreateAssessmentFrameworkControlSet

create_control_mapping_source

See CreateControlMappingSource

create_delegation_request

See CreateDelegationRequest

delegation

See Delegation

delegation_metadata

See DelegationMetadata

evidence

See Evidence

framework

See Framework

framework_metadata

See FrameworkMetadata

manual_evidence

See ManualEvidence

notification

See Notification

resource

See Resource

role

See Role

scope

See Scope

service_metadata

See ServiceMetadata

settings

See Settings

source_keyword

See SourceKeyword

update_assessment_framework_control_set

See UpdateAssessmentFrameworkControlSet

url

See Url

validation_exception_field

See ValidationExceptionField

Structs

Assessment

An entity that defines the scope of audit evidence collected by Audit Manager.

AssessmentControl

The control entity that represents a standard control or a custom control in an Audit Manager assessment.

AssessmentControlSet

Represents a set of controls in an Audit Manager assessment.

AssessmentEvidenceFolder

The folder where Audit Manager stores evidence for an assessment.

AssessmentFramework

The file used to structure and automate Audit Manager assessments for a given

AssessmentFrameworkMetadata

The metadata that's associated with a standard framework or a custom framework.

AssessmentFrameworkShareRequest

Represents a share request for a custom framework in Audit Manager.

AssessmentMetadata

The metadata that's associated with the specified assessment.

AssessmentMetadataItem

A metadata object that's associated with an assessment in Audit Manager.

AssessmentReport

A finalized document that's generated from an Audit Manager assessment. These

AssessmentReportEvidenceError

An error entity for the AssessmentReportEvidence API. This is used to

AssessmentReportMetadata

The metadata objects that are associated with the specified assessment report.

AssessmentReportsDestination

The location where Audit Manager saves assessment reports for the given

AwsAccount

The wrapper of Amazon Web Services account details, such as account ID or email address.

AwsService

An Amazon Web Service such as Amazon S3 or CloudTrail.

BatchCreateDelegationByAssessmentError

An error entity for the BatchCreateDelegationByAssessment API. This is

BatchDeleteDelegationByAssessmentError

An error entity for the BatchDeleteDelegationByAssessment API. This is

BatchImportEvidenceToAssessmentControlError

An error entity for the BatchImportEvidenceToAssessmentControl API. This

ChangeLog

The record of a change within Audit Manager. For example, this could be the

Control

A control in Audit Manager.

ControlComment

A comment that's posted by a user on a control. This includes the author's name, the

ControlMappingSource

The data source that determines where Audit Manager collects evidence from for

ControlMetadata

The metadata that's associated with the standard control or custom control.

ControlSet

A set of controls in Audit Manager.

CreateAssessmentFrameworkControl

The control entity attributes that uniquely identify an existing control to be added to

CreateAssessmentFrameworkControlSet

A controlSet entity that represents a collection of controls in Audit Manager. This doesn't contain the control set ID.

CreateControlMappingSource

The control mapping fields that represent the source for evidence collection, along

CreateDelegationRequest

A collection of attributes that's used to create a delegation for an assessment in

Delegation

The assignment of a control set to a delegate for review.

DelegationMetadata

The metadata that's associated with the delegation.

Evidence

A record that contains the information needed to demonstrate compliance with the

Framework

The file that's used to structure and automate Audit Manager assessments for a

FrameworkMetadata

The metadata of a framework, such as the name, ID, or description.

ManualEvidence

Evidence that's uploaded to Audit Manager manually.

Notification

The notification that informs a user of an update in Audit Manager. For

Resource

A system asset that's evaluated in an Audit Manager assessment.

Role

The wrapper that contains the Audit Manager role information of the current

Scope

The wrapper that contains the Amazon Web Services accounts and services that are in

ServiceMetadata

The metadata that's associated with the Amazon Web Service.

Settings

The settings object that holds all supported Audit Manager settings.

SourceKeyword

The keyword to search for in CloudTrail logs, Config rules,

UpdateAssessmentFrameworkControlSet

A controlSet entity that represents a collection of controls in Audit Manager. This doesn't contain the control set ID.

Url

Short for uniform resource locator. A URL is used as a unique identifier to locate a

ValidationExceptionField

Indicates that the request has invalid or missing parameters for the field.

Enums

AccountStatus

ActionEnum

AssessmentReportDestinationType

AssessmentReportStatus

AssessmentStatus

ControlResponse

ControlSetStatus

ControlStatus

ControlType

DelegationStatus

FrameworkType

KeywordInputType

ObjectTypeEnum

RoleType

SettingAttribute

ShareRequestAction

ShareRequestStatus

ShareRequestType

SourceFrequency

SourceSetUpOption

SourceType

ValidationExceptionReason